Data Driven Nonparametric Test for Stochastic Ordering

Performs the data driven smooth test for detection of the stochastic dominance of the first order, as described in detail in Nonparametric tests for stochastic ordering Ledwina, T. and Wyłupek, G. TEST (2012) 21: 730. https://doi.org/10.1007/s11749-011-0278-7

ddst.stochasticorder.test(x, y, t = 2.2, k.N = 4)

Arguments

x
a (non-empty) numeric vector of data values

y
a (non-empty) numeric vector of data values

t
tuning parameter, see test description

k.N
The choice of k(N) belongs to the user and should reflect the individual aims and needs. Some authors’ recommendation is given in Section 5.2.

Details

It has been proved that the error of the first kind of the test based on QT attains the maximal value under F = G. Moreover, under F = G the statistic QT is distribution-free. Hence, it is enough to calculate the critical values and p-values under an arbitrary continuous distribution functions F = G.

Examples

set.seed(13)
library("rmutil")
# 1. Pareto(1)/Pareto(1.5)
#uzyc parametrow z tabeli 3, p. 742, np dla m = n = 50,
x = rpareto(50, 1, 1.5)
y = rpareto(50, 1, 1.5)
ddst.stochasticorder.test(x, y)

#> 
#> 	Data Driven Nonparametric Test for Stochastic Ordering
#> 
#> data:  
#> stat.T = 2.56, stat.S = 2.56, ncoord.T = 1, ncoord.S = 1
#> 

# 2. Laplace(0,1)/Laplace(1,25)
x = rlaplace(50, 0, 1)
y = rlaplace(50, 1, 25)
ddst.stochasticorder.test(x, y)

#> 
#> 	Data Driven Nonparametric Test for Stochastic Ordering
#> 
#> data:  
#> stat.T = 111.11, stat.S = 12.00, ncoord.T = 5, ncoord.S = 2
#> 

# 3. LN(0.85,0.6)/LN(1.2,0.2)
x = rlnorm(50, 0.85, 0.6)
y = rlnorm(50, 1.2, 0.2)
ddst.stochasticorder.test(x, y)

#> 
#> 	Data Driven Nonparametric Test for Stochastic Ordering
#> 
#> data:  
#> stat.T = 0, stat.S = 0, ncoord.T = 1, ncoord.S = 1
#> 

# NOT RUN {
# Generate distribution of test statistic
N <- 1000
samp <- replicate(N, {
   x = runif(30)
   y = runif(30)
   # statistics with Schwartz penalty
   ddst.stochasticorder.test(x, y)$statistic[2]
})
quantile(samp, 0.95)
plot(ecdf(samp))
# }