Data Driven Test for k-Sample Umbrella Alternatives

Performs data driven smooth test for so-called umbrella alternatives in k-sample problem; see Wylupek (2016).

ddst.umbrella.test(x, ..., tlh.p = 2.2, tl.n = 2.2, p = 3,
  type = "T", B = 1000, compute.p = FALSE)

Arguments

x
a list with vectors or vectors (in thos case other vectors will be in ...)

...
other vectors

B
an integer specifying the number of replicates used in p-value computation

compute.p
a logical value indicating whether to compute a p-value

d_N
an integer, number of coordinates, see package description

c
a positive number, penalty for model selection rule, see package description

Examples

# H0 is true
x = runif(80)
y = runif(80)
z = runif(80)
ddst.umbrella.test(x, y, z, compute.p=TRUE)

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> T = 0, l = NA, p-value = 0.719
#> 

ddst.umbrella.test(x, y, z, compute.p=TRUE, type = "S")

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> S = 0, l = NA, p-value = 0.719
#> 

ddst.umbrella.test(x, y, z, compute.p=TRUE, type = "M")

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> M = -12.529, l = NA, p-value = 0.727
#> 

ddst.umbrella.test(list(x, y, z), compute.p=TRUE)

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> T = 0, l = NA, p-value = 0.705
#> 

# known fixed alternative
x = runif(80)
y = rbeta(80,4,2)
w = rnorm(30)
z = rexp(10, 1)
ddst.umbrella.test(x, y, w, z, compute.p=TRUE)

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> T = -170.62, l = NA, p-value = 1
#> 

ddst.umbrella.test(x, y, w, z, compute.p=TRUE, type = "S")

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> S = -170.62, l = NA, p-value = 1
#> 

ddst.umbrella.test(x, y, w, z, compute.p=TRUE, type = "M")

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> M = -190.91, l = NA, p-value = 1
#> 

ddst.umbrella.test(list(x, y, w, z), compute.p=TRUE)

#> 
#> 	Data Driven k-Sample Umbrella Test
#> 
#> data:  
#> T = -170.62, l = NA, p-value = 1
#>